Switch to Nextcloud; implement HTTPS on local network
I wasn't very happy with Filerun's closed-source nature.
Because my router doesn't support Hairpin NAT (NAT Loopback),
I need to be able to locally access Nextcloud, Immich, and
any other service that I can't proxy through Cloudflare.
When I'm not on LAN, I can access these services through
Wireguard.

Originally, I planned to just use HTTP as I had previously
done for the -arr stack, but the annoying warnings and the
idea of unencrypted transfers on my LAN prompted me to
implement HTTPS through Caddy. It was as simple as setting
"tls internal" for those addresses and accepting the self-
signed root certificate on my devices. Because it was so
easy, I decided to proxy the -arr stack as well!